extract fake users to separate gateway
