+++ client_dashboard/desktop/src/jvmMain/kotlin/di/GatewayModule.kt
@@ -6,9 +6,11 @@ import org.koin.dsl.module
 import org.thechance.common.data.local.gateway.IdentityGateway
 import org.thechance.common.data.local.gateway.LocalGateway
 import org.thechance.common.data.remote.gateway.FakeRemoteGateway
+import org.thechance.common.data.remote.gateway.UsersGateway
 import org.thechance.common.domain.getway.IIdentityGateway
 import org.thechance.common.domain.getway.ILocalGateway
 import org.thechance.common.domain.getway.IRemoteGateway
+import org.thechance.common.domain.getway.IUsersGateway
 
 val GatewayModule = module {
     //TODO remove comment when finish testing
@@ -16,4 +18,6 @@ val GatewayModule = module {
     singleOf(::LocalGateway) { bind<ILocalGateway>() }
     singleOf(::FakeRemoteGateway) { bind<IRemoteGateway>() }
     singleOf(::IdentityGateway) { bind<IIdentityGateway>() }
+    singleOf(::UsersGateway) { bind<IUsersGateway>() }
+//    singleOf(::UsersFakeGateway){ bind<IUsersGateway>()}
 }
